requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
geoalchemy2
python-dotenv
pyyaml
spacy
gensim
scikit-learn
pandas
numpy
unidecode